Fix bug: The value of the program counter is one larger.
VariableDeclaration::getLocalVariableName method needs the value of the program counter.
bytes.getIndex() is used for the program counter.
But, bytes.getIndex() is one greater than the value of the program counter.

Pass1::parseInstruction method executes the following.

		int currentIndex= bytes.getIndex();
		short opcode= (short) bytes.readUnsignedByte();

currentIndex is the program counter.
bytes.readUnsignedByte() increment bytes.getIndex().

This commit changes the value used for the program counter from bytes.getIndex() to currentIndex.

In addition, with this commit, LocalVariableTable will not be used.

For example, suppose we write the following method.

    void test() {
        String value = "100";
        Integer.valueOf(value);
    }

The Java compiler generates the following bytecode.

    0 ldc #2 <100>
    2 astore_1
    3 aload_1
    4 invokestatic #3 <java/lang/Integer.valueOf>
    7 pop
    8 return

    LocalVariableTable:
    0 0 10 0 this Lsample/Test;
    1 3 10 1 value Ljava/lang/String;

In the LocalVariableTable entry of 'value', start_pc=3 and length=10.
'astore' is not included in the scope.
The getLocalVariableName method returns 'l1' for 'astore' and 'value' for 'aload'.

This problem turned out as the Kotlin compiler inserted 'nop' between 'asotre' and 'aload'.
It was not found because the value of the program counter was misaligned.